requirements.txt
Django==1.9.7
django-filter==0.13.0
djangorestframework==3.4.0
Markdown==2.6.6
mysqlclient==1.3.7
oauthlib==1.1.2
PyJWT==1.4.1
python-openid==2.2.5
python-social-auth==0.2.19
requests==2.10.0
requests-oauthlib==0.6.2
six==1.10.0
Flask==0.11.1
django-cors-headers
